requirements.txt
anthropic>=0.35.0
PyYAML
jsonschema
psutil
